requirements.txt
streamlit==1.46.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
sentence-transformers==2.5.1
google-api-python-client==2.123.0
python-dotenv==1.0.1
google-generativeai==0.4.1
